Fixed casted-lvalue warnings/errors for GCC 4+
The latest versions of GCC do not like casted-lvalues assignments like:

((int)i) = 5;

To fix this, the WAM engine PCell union was extended to include backward-compatibility types. This eliminated the need for casting in many cases. In a few additional places, castes were removed or shifted, or explicit lvalue macros were defined.