Adjust echo to be router
The original echo script could not echo file or path-like requests such as "/abc.xyz" or "/abc/xyz", because the default PHP dev router intercepted them and 404-ed.
When used as a router script, all request are handled by echo, so there are no limits on request path echoing. An additional benefit is the quieting of log output.