blt_shlc.pro: Changed get_command_line_info to preserve value of ImageName in CLInfo.
blt_pckg.pro: - Modified attach_image0/2 to obtain the (starting) image name, and if that image name is not one of (path to) alspro, alsdev, alspro.exe, alsdev.exe (or UC on last 2), then it does not concern itself with saving command_line/1.  When the starting image is not one of those, this concern was a bug blocking the creation of (yet another) saved image “on top of” the current starting image.
- blt_pckg.pro: Added get_app_cmd_line/4 to allow saved images to easily obtain and process their command lines.

Added ~tests/tsuite/test_app_image.sh to test both a "single-level" saved image together with a "
double-level"  saved image which creates a saved image "on top of" an existing saved image. Add ~
tests/tsuite/savedimages/app_image_test[0-2].pro for use by this, and modified ~generic/gnu_makefile:testrun to invoke test_app_image.sh.