Bltobp (#152) - Suppress generation of builtin OBP files
Fixes #92.  Suppress generation of builtin OBPs for consistency.

Original commit messages combined in this commit:

* @kenbowen Calls to ‘$load’(___,0) in builtins.pro go to generic/loadfile.c:load_file.
Changing the 0 to 2, makes options==2 in load_file.  The modifications in load_file suppress .obp creation when options==2, but allow the rest of normal file loading to proceed ok.  The autoload call (to load builtins.pro itself) in main.c is changed similarly.

* @chuckhoupt Refactor to use bitmask for options and removed debug code.